Extract methods to collapse votes by user into DarCollectionUtils
