+++ src/components/collection_voting_slab/MultiDatasetVoteSlab.js
@@ -9,6 +9,7 @@ import DatasetsRequestedPanel from "./DatasetsRequestedPanel";
 import {dataUsePills} from "./ResearchProposalVoteSlab";
 import {formatDate} from "../../libs/utils";
 import {
+  collapseVotesByUser,
   extractDacUserVotesFromBucket,
   extractDatasetIdsFromBucket,
   extractUserVotesFromBucket
@@ -49,20 +50,6 @@ const styles = {
     flexDirection: 'column',
     rowGap: '1.5rem'
   }
-};
-
-const addIfUnique = (newValue, existingValues) => {
-  if(!isNil(newValue) && !includes(newValue, existingValues)) {
-    existingValues.push(newValue);
-  }
-};
-
-const appendAll = (values) => {
-  let result = '';
-  forEach(value => {
-    result += `${value}\n`;
-  })(values);
-  return !isEmpty(result) ? result : null;
 };
 
 export default function MultiDatasetVoteSlab(props) {
@@ -106,60 +93,16 @@ export default function MultiDatasetVoteSlab(props) {
   };
 
   const ChairVoteInfo = () => {
-    const votesGroupedByUser = groupBy(vote => vote.dacUserId)(cloneDeep(dacVotes));
-    const votesCollapsedByUser = flatMap(userIdKey => {
-      const votesByUser = votesGroupedByUser[userIdKey];
-      const collapsedVotes = collapseVotes({votes: votesByUser});
-      return convertToVoteObjects({collapsedVotes});
-    })(Object.keys(votesGroupedByUser));
-
     return div({style: styles.chairVoteInfo, isRendered: isChair && dacVotes.length > 0, dataCy: 'chair-vote-info'}, [
       h(VotesPieChart, {
         votes: dacVotes,
       }),
       div(['My DAC\'s Votes (detail)']),
       h(VoteSummaryTable, {
-        dacVotes: votesCollapsedByUser,
+        dacVotes: collapseVotesByUser(dacVotes),
         isLoading,
       })
     ]);
-  };
-
-  const collapseVotes = ({votes}) => {
-    const collapsedVotes = {};
-    forEach( vote => {
-      const matchingVote = collapsedVotes[`${vote.vote}`];
-      if (isNil(matchingVote)) {
-        collapsedVotes[`${vote.vote}`] = {
-          vote: vote.vote,
-          voteId: vote.voteId,
-          displayName: vote.displayName,
-          rationales: !isNil(vote.rationale) ? [vote.rationale] : [],
-          createDates: !isNil(vote.createDate) ? [vote.createDate] : []
-        };
-      }
-      else {
-        addIfUnique(vote.rationale, matchingVote.rationales);
-        addIfUnique(vote.createDate, matchingVote.createDates);
-      }
-    })(votes);
-    return collapsedVotes;
-  };
-
-  const convertToVoteObjects = ({collapsedVotes}) => {
-    return map( key => {
-      const collapsedVote = collapsedVotes[key];
-      const collapsedRationale = appendAll(collapsedVote.rationales);
-      const collapsedDate = appendAll(map(date => formatDate(date))(collapsedVote.createDates));
-
-      return {
-        vote: collapsedVote.vote ,
-        voteId: collapsedVote.voteId,
-        displayName: collapsedVote.displayName,
-        rationale: collapsedRationale,
-        createDate: collapsedDate
-      };
-    })(Object.keys(collapsedVotes));
   };
 
   const DatasetsRequested = () => {
