+++ src/utils/DarCollectionUtils.js
@@ -1,6 +1,24 @@
-import { flow, isEmpty, map, join, filter, forEach, flatMap, toLower, sortBy, isNil, size, includes, get, concat, findIndex, cloneDeep } from 'lodash/fp';
+import {
+  flow,
+  isEmpty,
+  map,
+  join,
+  filter,
+  forEach,
+  flatMap,
+  toLower,
+  sortBy,
+  isNil,
+  size,
+  includes,
+  get,
+  concat,
+  findIndex,
+  cloneDeep,
+  groupBy
+} from 'lodash/fp';
 import {translateDataUseRestrictionsFromDataUseArray} from '../libs/dataUseTranslation';
-import { Notifications } from '../libs/utils';
+import {formatDate, Notifications} from '../libs/utils';
 import { Collections } from '../libs/ajax';
 
 //Initial step, organizes raw data for further processing in later function/steps
@@ -199,6 +217,69 @@ export const extractDatasetIdsFromBucket = (bucket) => {
   )(bucket);
 };
 
+//collapses votes by the same user with same vote (true/false) into a singular vote with appended rationales / dates if different
+export const collapseVotesByUser = (votes) => {
+  const votesGroupedByUser = groupBy(vote => vote.dacUserId)(cloneDeep(votes));
+  return flatMap(userIdKey => {
+    const votesByUser = votesGroupedByUser[userIdKey];
+    const collapsedVotes = collapseVotes({votes: votesByUser});
+    return convertToVoteObjects({collapsedVotes});
+  })(Object.keys(votesGroupedByUser));
+};
+
+//helper method to collapse votes that converts them to an object with differing rationales and dates in arrays
+const collapseVotes = ({votes}) => {
+  const collapsedVotes = {};
+  forEach( vote => {
+    const matchingVote = collapsedVotes[`${vote.vote}`];
+    if (isNil(matchingVote)) {
+      collapsedVotes[`${vote.vote}`] = {
+        vote: vote.vote,
+        voteId: vote.voteId,
+        displayName: vote.displayName,
+        rationales: !isNil(vote.rationale) ? [vote.rationale] : [],
+        createDates: !isNil(vote.createDate) ? [vote.createDate] : []
+      };
+    }
+    else {
+      addIfUnique(vote.rationale, matchingVote.rationales);
+      addIfUnique(vote.createDate, matchingVote.createDates);
+    }
+  })(votes);
+  return collapsedVotes;
+};
+
+//helper method to follow collapseVotes in flow that creates standard vote objects
+const convertToVoteObjects = ({collapsedVotes}) => {
+  return map( key => {
+    const collapsedVote = collapsedVotes[key];
+    const collapsedRationale = appendAll(collapsedVote.rationales);
+    const collapsedDate = appendAll(map(date => formatDate(date))(collapsedVote.createDates));
+
+    return {
+      vote: collapsedVote.vote ,
+      voteId: collapsedVote.voteId,
+      displayName: collapsedVote.displayName,
+      rationale: collapsedRationale,
+      createDate: collapsedDate
+    };
+  })(Object.keys(collapsedVotes));
+};
+
+const appendAll = (values) => {
+  let result = '';
+  forEach(value => {
+    result += `${value}\n`;
+  })(values);
+  return !isEmpty(result) ? result : null;
+};
+
+const addIfUnique = (newValue, existingValues) => {
+  if(!isNil(newValue) && !includes(newValue, existingValues)) {
+    existingValues.push(newValue);
+  }
+};
+
 //Admin only helper function
 export const checkIfOpenableElectionPresent = (dars) => {
   const darCount = size(dars);
@@ -276,5 +357,6 @@ export default {
   processDataUseBuckets,
   extractDacUserVotesFromBucket,
   extractUserVotesFromBucket,
-  extractDatasetIdsFromBucket
+  extractDatasetIdsFromBucket,
+  collapseVotesByUser
 };
